requirements.txt
asgiref==3.4.1
cycler==0.10.0
dataclasses-json==0.5.5
diversipy==0.8
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
doepy==0.0.1
kiwisolver==1.3.2
marshmallow==3.13.0
marshmallow-enum==1.5.1
matplotlib==3.4.3
mypy-extensions==0.4.3
numpy==1.21.2
numpy-stl==2.16.3
pandas==1.3.2
Pillow==8.3.2
plot3d==1.3.1
psutil==5.8.0
psycopg2==2.9.1
pyDOE==0.3.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.1
scipy==1.7.1
six==1.16.0
sqlparse==0.4.1
tqdm==4.62.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
